refactor: extract controlpanel hooks and remove unused deps (#81, #171)
- extract usePresetOperations hook (preset CRUD, dirty detection, dashboard layout)
- extract useProfileImportExport hook (project export/import with Zod validation)
- reduce ControlPanel from 528 to 306 lines
- remove unused typed-emitter and eventemitter3 packages (#171)
- fix pre-existing set-state-in-effect lint error in port refresh

via [Happy](https://happy.engineering)
