+++ package.json
@@ -46,7 +46,6 @@
     "ansi-to-react": "6.1.6",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "eventemitter3": "^5.0.4",
     "cmdk": "^1.1.1",
     "i18next": "23.10.0",
     "immer": "^11.1.3",
@@ -64,7 +63,6 @@
     "sonner": "^2.0.7",
     "tailwind-merge": "^2.6.0",
     "ts-pattern": "^5.9.0",
-    "typed-emitter": "^2.1.0",
     "zod": "3.25.76",
     "zod-validation-error": "^5.0.0",
     "zustand": "^5.0.10"
